Add seq counting to prod/cons
To mitigate a race condition where a producer or consumer successfully
advances the head using a stale view that happens to match the current
state well enough to pass the compare exchange on the head. Very small
queues are prone to this problem.

Note this doesn't entirely prevent the race condition, but makes it
much, much less likely to occur. Need to think on this more...